refactor(eslint-plugin-react-x): move purity and refs resolvers to lib.ts
